add support for lottie light player
