+++ example/src/Test.js
@@ -1,4 +1,5 @@
 import Lottie from 'react-lottie-player'
+import LottieLight from 'react-lottie-player/dist/LottiePlayerLight'
 
 import React from 'react';
 
@@ -16,6 +17,22 @@ const Test = () => {
 
     case '/test/2': return (
       <Lottie
+        goTo={60}
+        path="/26514-check-success-animation.json"
+        style={{ width: 150, height: 150 }}
+      />
+    );
+
+    case '/test/3': return (
+      <LottieLight
+        goTo={60}
+        animationData={lottieJson}
+        style={{ width: 150, height: 150 }}
+      />
+    );
+
+    case '/test/4': return (
+      <LottieLight
         goTo={60}
         path="/26514-check-success-animation.json"
         style={{ width: 150, height: 150 }}
