+++ src/LottiePlayer.js
@@ -1,0 +1,4 @@
+import lottie from 'lottie-web'
+import makeLottiePlayer from './makeLottiePlayer'
+
+export default makeLottiePlayer(lottie)
