+++ src/LottiePlayer.test.js
