+++ src/LottiePlayerLight.js
@@ -1,0 +1,4 @@
+import lottie from 'lottie-web/build/player/lottie_light'
+import makeLottiePlayer from './makeLottiePlayer'
+
+export default makeLottiePlayer(lottie)
